+++ tests/conftest.py
@@ -81,6 +81,7 @@ _SAMPLE_VOCAB_DIR = _SAMPLE_DATA_DIR / "distilbert-tokenizer-files"
 
 CHECKPOINT_PATH = str(_SAMPLE_VOCAB_DIR)
 MODEL_CFG = {
+    "name": "Default Model",
     "model": {
         "class_name": "tests.test_loading_resources.load_hf_text_classif_pipeline",
         "kwargs": {"checkpoint_path": CHECKPOINT_PATH},
@@ -88,6 +89,7 @@ MODEL_CFG = {
 }
 
 HIGH_THRESHOLD_MODEL = {
+    "name": "High threshold Model",
     "model": {
         "class_name": "tests.test_loading_resources.load_hf_text_classif_pipeline",
         "kwargs": {"checkpoint_path": CHECKPOINT_PATH},
@@ -111,6 +113,38 @@ def simple_text_config(tmp_path):
         name="sentiment-analysis",
         dataset=DATASET_CFG,
         pipelines=[MODEL_CFG],
+        artifact_path=str(tmp_path),
+        batch_size=10,
+        use_cuda="auto",
+        model_contract="hf_text_classification",
+        saliency_layer="distilbert.embeddings.word_embeddings",
+        rejection_class=None,
+        behavioral_testing=SIMPLE_PERTURBATION_TESTING_CONFIG,
+    )
+
+
+@pytest.fixture
+def simple_multipipeline_text_config(tmp_path):
+    return AzimuthConfig(
+        name="sentiment-analysis",
+        dataset=DATASET_CFG,
+        pipelines=[MODEL_CFG, HIGH_THRESHOLD_MODEL],
+        artifact_path=str(tmp_path),
+        batch_size=10,
+        use_cuda="auto",
+        model_contract="hf_text_classification",
+        saliency_layer="distilbert.embeddings.word_embeddings",
+        rejection_class=None,
+        behavioral_testing=SIMPLE_PERTURBATION_TESTING_CONFIG,
+    )
+
+
+@pytest.fixture
+def simple_no_pipeline_text_config(tmp_path):
+    return AzimuthConfig(
+        name="sentiment-analysis",
+        dataset=DATASET_CFG,
+        pipelines=None,
         artifact_path=str(tmp_path),
         batch_size=10,
         use_cuda="auto",